+++ src/main/java/com/github/unchama/util/failable/Try.java
@@ -1,5 +1,6 @@
-package com.github.unchama.util;
+package com.github.unchama.util.failable;
 
+import com.github.unchama.util.ActionStatus;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.Arrays;
