+++ src/main/java/com/github/unchama/util/failable/TryWithoutFailValue.java
@@ -1,5 +1,7 @@
-package com.github.unchama.util;
+package com.github.unchama.util.failable;
 
+import com.github.unchama.util.ActionStatus;
+import com.github.unchama.util.Unit;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.Arrays;
